Move BackupTestHelper class into separate file
